requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.40
chromadb>=0.5.0
sentence-transformers>=2.2.0,<3.0.0
rank-bm25>=0.2.2
pydantic>=2.8.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
yfinance>=0.2.40
feedparser>=6.0.10
streamlit>=1.39.0
rich>=13.7.0
tabulate>=0.9.0
numpy>=1.26.0,<2.0
pandas>=2.2.0
